v0.1.3

+++ src/gemini/main.py
@@ -3,41 +3,16 @@ import json
 import time
 
 from tqdm import tqdm
-from dotenv import load_dotenv
+# from dotenv import load_dotenv
 
 import google.generativeai as genai
-from google.oauth2.credentials import Credentials
-from google.auth.transport.requests import Request
-from google_auth_oauthlib.flow import InstalledAppFlow
+# from google.oauth2.credentials import Credentials
+# from google.auth.transport.requests import Request
+# from google_auth_oauthlib.flow import InstalledAppFlow
 
-load_dotenv()
+from load_creds import load_creds
 
-SCOPES = ["https://www.googleapis.com/auth/cloud-platform", "https://www.googleapis.com/auth/generative-language.tuning"]
-KEY_FILE = os.getenv("PARENT_DIR") + "auth/credentials.json"
-TOKEN_FILE = os.getenv("PARENT_DIR") + "auth/token.json"
-
-# auth stuff, ugh
-creds = None
-
-if os.path.exists(TOKEN_FILE):
-    creds = Credentials.from_authorized_user_file(TOKEN_FILE, SCOPES)
-
-if not creds or not creds.valid:
-    if creds and creds.expired and creds.refresh_token:
-        creds.refresh(Request())
-    else:
-        flow = InstalledAppFlow.from_client_secrets_file(KEY_FILE, SCOPES)
-        creds = flow.run_local_server()
-    
-    with open(TOKEN_FILE, "w") as token:
-        token.write(creds.to_json())
-
-genai.configure(credentials=creds)        
-
-# code starts here
-
-# for model_info in genai.list_tuned_models():
-#     print(model_info)
+load_creds()
 
 name = "test-model-5"
 
@@ -46,8 +21,9 @@ model = genai.GenerativeModel(model_name=f"tunedModels/{name}")
 with open("./data/final_data.json", "r") as f:
     data = json.loads(f.read())
 
-with open(f"./test/{name}.txt", "w+") as f:
-    f.write("")
+if not os.path.exists(f"./test/{name}.txt"):
+    with open(f"./test/{name}.txt", "w+") as f:
+        f.write("")
 
 for question, answer in tqdm(data[54:]):
     try:
